Move index.js to src
